requirement.txt
# 环境所需依赖

django==3.2.2
scrapy==2.4.1
mysqlclient==2.0.3
scrapy-redis==0.6.8
beautifulsoup4==4.9.3
requests==2.25.1